basic test for -d, --rdns
Usually, -d gives the same result as -n.  Test this.

When the target is a name that resolves to an IP, but the IP
does not resolve to the same name or any name at all, the
output differs.  This is harder to test, because this is
usually seen as a misconfiguration, so leave this out for now.